Build LLVM/clang libs during SANITIZE_HOST build
In a regular build, the checked-in prebuilts of this library
(prebuilts/sdk/tools/linux) are used.  They are rebuilt only when
FORCE_BUILD_LLVM_COMPONENTS is set.  However, they also need to be built
during host sanitize builds since we don't have sanitized prebuilts for
these libraries.  This change forces the libraries to be built if
SANITIZE_HOST is set.

Test: Verify expected behavior during:
      m;
      m SANITIZE_HOST=address;
      m FORCE_BUILD_LLVM_COMPONENTS=true;

Change-Id: I81ba1ad0f1a800a3ee7c1d36eba60f9c8e5dc8c9