linux_bionic_arm64 is always built from source
Previously, libLLVM_android which is inheriting from
force_build_llvm_components_defaults, was disabled for all host targets
unless FORCE_BUILD_LLVM_COMPONENTS is set to true. The implicit
assumption here was that we will use prebuilts in prebuilts/sdk/tools
for hosts.

But this won't work as we add arm64 support to the Linux_bionic target.
We don't have prebuilts for the target, and even if we had it, it
doesn't look ideal to use the prebuilt because we always have been
building the library for the ARM architecture. We can use the same for
the new host target as well.

In order to not rely on the existence of prebuilts, the source-build
libLLVM_android is kept enabled for the linux_bionic_arm64 target.

Bug: 159685774
Test: HOST_CROSS_OS=linux_bionic HOST_CROSS_ARCH=arm64 m

Change-Id: I4e4eff6c27ea86b8e98d2de3732755173a6731fe